requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose
argon2-cffi
redis
dotenv
passlib
email-validator